Bugfix - colliding indexes
It turns out that the data for each beautiful coffee site is shared in the browser. Therefore, the "count" variable was being overwritten across applications, which could cause applications to error or overwrite progress on applications that they should not have. For example, there was a bug where NLVE would write an index of 2000, which would translate into an index of 2000 on Aeneid, causing the app to display a blank screen, which is not ideal, even though the error is easily fixable by just clicking "next". To fix collisions, each application has been fitted with application specific indexes.